Address mypy issues

+++ ote_sdk/ote_sdk/entities/train_parameters.py
@@ -29,7 +29,7 @@ class UpdateProgressCallback(Protocol):
 
 
 # pylint: disable=unused-argument
-def default_progress_callback(progress: int, score: Optional[float] = None):
+def default_progress_callback(progress: float, score: Optional[float] = None):
     """
     Default progress callback. It is a placeholder (does nothing) and is used in empty TrainParameters.
     """
